Refactor RichParameterInput stories and flat MultiTextField
